Add rebuild change set index route
This commit adds the rebuild change set index route. There is no
corresponding button or action that can trigger it in the frontend yet,
but it is available.

This commit also standardizes the two rebuild functions' trace naming
schemes and moves them to the root index module. In addition to these
changes, all routes are now in their own submodules.
