+++ lib/sdf-server/src/service/v2/index/get_workspace_index.rs
@@ -1,0 +1,27 @@
+use std::collections::HashMap;
+
+use axum::{extract::Path, Json};
+
+use dal::{ChangeSet, ChangeSetId, WorkspacePk};
+use si_frontend_types::object::FrontendObject;
+
+use crate::extract::{FriggStore, HandlerContext};
+
+use super::{AccessBuilder, IndexResult};
+
+pub async fn get_workspace_index(
+    HandlerContext(builder): HandlerContext,
+    AccessBuilder(access_builder): AccessBuilder,
+    FriggStore(frigg): FriggStore,
+    Path(workspace_pk): Path<WorkspacePk>,
+) -> IndexResult<Json<HashMap<ChangeSetId, Option<FrontendObject>>>> {
+    let ctx = builder.build_head(access_builder).await?;
+
+    let mut indexes = HashMap::new();
+    for change_set in ChangeSet::list_active(&ctx).await? {
+        let maybe_index = frigg.get_index(workspace_pk, change_set.id).await?;
+        indexes.insert(change_set.id, maybe_index.map(|i| i.0));
+    }
+
+    Ok(Json(indexes))
+}
