+++ lib/sdf-server/src/service/v2/index/rebuild_change_set_index.rs
@@ -1,0 +1,13 @@
+use axum::extract::Path;
+use dal::{ChangeSetId, WorkspacePk};
+
+use crate::extract::EddaClient;
+
+use super::{request_rebuild, IndexResult};
+
+pub async fn rebuild_change_set_index(
+    EddaClient(edda_client): EddaClient,
+    Path((workspace_pk, change_set_id)): Path<(WorkspacePk, ChangeSetId)>,
+) -> IndexResult<()> {
+    request_rebuild(&edda_client, workspace_pk, change_set_id).await
+}
